Fix UIPageControl on RootPageViewController
The fix I implemented is not ideal. It uses transform to change the
size of the dots in the UIPageControl. I do not know if this way
it the correct solution but it will be enough for now.

Unfortunately, all the constraints seem to be behaving in a unexpected
manner in this screen.