Fix InboxViewController's UICollectionView
Motivation:
The InboxViewController's UICollectionView was using an older API for
building its layout. This resulted in awkward layout to build and
maintain. Also, the InboxCollectionViewCell wasn't yet implemented
according to the design specs.

Modification:
* I have reworked the layout of the UICollectionView using a more recent
API, UICollectionViewCompositionalLayout.
* I have implemented the InboxCollectionViewCell following the
prototype. It has a new function called 'configure' that should be used
to configure the cell from the UICollectionViewDataSource.

Result:
* The new InboxViewController's UICollectionView is easier to maintain
and understand.
* The InboxCollectionViewCell has been implemented according to the
prototype.